// src/app/core/models/itaskeen.ts
export interface Itaskeen {
    clientName: string
    menCount: number
    womenCount: number
    totalCount: number
    hotelName: string
    bedsCountInRoom: number
    // numOfRooms: number
    reservedRooms: Room_obj[]
    num_of_all_rooms: number
    numOfSingleBeds: number
    fromDate: string
    toDate: string
    stayType: number
    gender: number
    invoiceNumber: string
    employeeName: string
    paymentMethod: number
}
// ...
export interface Room_obj {
    type: number;
    roomsCount: number;
}